refactor(challenge): BonusMissionModal을 실사용처 my-challenge/modal/로 이동
P1에서 이름만 보고 bonus-mission/modal/에 뒀으나, 실제 사용처는
my-challenge/section/mission-submit 뿐 → "배치는 참조가 결정한다" 원칙에 맞게 이동.
